requirements.txt
langchain
openai
faiss-cpu
streamlit
tiktoken
python-dotenv
